EPoA scraper to be able to use PDFs from URLs
